[msm_shared/crypto]: Fix error in hash calculation
For data size which is not a multiple of 4 bytes, we calculated the
hash for the left over bytes by wrong byte order. Now we change the
byte order for leftover bytes and calculate the hash properly.

This is a corner case as we always verified images with size of
multiple of 4 bytes.

(cherry picked from commit 22d53c65e803468c4fc10f04ea6b91fccb36642b)

Change-Id: Ieafbc34ac478e8c6e90d2412834bad69a902b55f