aboot/sparse: Fix default case and flashing for sparse parsing
For don't care blocks the total_block needs to be updated as
this is used for the offset to write on emmc/mmc. Default
case is also missing the default tag and never enters
when an unrecognized type is detected.

(cherry picked from commit da29b1e973ce8b815dbf32ff83d48f2f1aae7f37)

Change-Id: I2cb04d8f3991da8e562b73dbc2097707cbd8dee7